+++ Shared/Support/Participant+Helpers.swift
@@ -14,13 +14,7 @@
  * limitations under the License.
  */
 
-import AVFoundation
 import LiveKit
-import SwiftUI
-
-import CoreImage.CIFilterBuiltins
-import ReplayKit
-import WebRTC
 
 public extension Participant {
     var mainVideoPublication: TrackPublication? {
@@ -33,27 +27,5 @@ public extension Participant {
 
     var subVideoTrack: VideoTrack? {
         firstScreenShareVideoTrack != nil ? firstCameraVideoTrack : nil
-    }
-}
-
-struct ExampleRoomMessage: Identifiable, Equatable, Hashable, Codable {
-    // Identifiable protocol needs param named id
-    var id: String {
-        messageId
-    }
-
-    // message id
-    let messageId: String
-
-    let senderSid: String
-    let senderIdentity: String?
-    let text: String
-
-    static func == (lhs: Self, rhs: Self) -> Bool {
-        lhs.messageId == rhs.messageId
-    }
-
-    func hash(into hasher: inout Hasher) {
-        hasher.combine(messageId)
     }
 }
